FuzzySystems.FuzzyLogic.@rule — Macro

@rule(expr...)

Create a fuzzy Rule.

Examples

julia> @rule cheap tip = poor service & bad food
Rule((:poor, :bad), :cheap, "MAX")

FuzzySystems.FuzzyLogic.@rules — Macro

@rules(exprs...)

Prepare multiple fuzzy rules of the form Consequent = Antecedent₁ [&, |] Antecedent₂ ...

Examples

julia> @rules {
    cheap tip = poor service & bad food
    average tip = good service
}
# shorthand form
julia> @rules {
    cheap = poor & bad
    average = good
}

FuzzySystems.FuzzyLogic.@var — Macro

@var(exprs...)

Syntax shortcut for making membership dictionaries.

Examples

julia> @var {
    poor     : Q 0 0 2 4
    good     : T 3 5 7
}
Dict{Symbol, MF} with 3 entries:
    :excellent => Q|₆₌⁸⁼¹⁼₁₌|
    :good => T|₃₌⁵⁼₇₌|

FuzzySystems.FuzzyLogic.implicationproperties — Method

Axiom adherence of an implication function

N is a strong negation function, i.e. a fuzzy complement ~~x == x. T is a left-continuous t-norm, defaults to nilpotent minimum

Mas, M., Monserrat, M., Torrens, J., & Trillas, E. (2007). A survey on fuzzy implication functions. IEEE Transactions on fuzzy systems, 15(6), 1107-1121.

FuzzySystems.FuzzyLogic.issnorm — Method

is a fuzzy union (s-norm, AND) of the form [0,1]² -> [0,1]

associative, 2) monotone, 3) communicative and 4) bounded?

FuzzySystems.FuzzyLogic.istnorm — Method

Is a fuzzy intersection (t-norm, OR) of the form [0,1]² -> [0,1]

associative, 2) monotone, 3) communicative and 4) bounded?

FuzzySystems.FuzzyLogic.mf — Method

mf(s::Symbol, args...)

Construct a membership function using an alias.

Example

julia> mf(:G, 0, 1) # gaussian membership
G|⁰⁼⁄₁₌|

FuzzySystems.FuzzyLogic.μ — Function

μ(x, mf::membership_function)

Obtain degrees of membership $μ$ for a given crisp input $x$

$μ_{triangular}(x)= max(min(\frac{x\,-\,l}{t\,-\,l}, \frac{r\,-\,x}{t\,-\,r}), 0)$, on left, top and right vertices
$μ_{trapezoid}(x)= max(min(\frac{x\,-\,lb}{lt\,-\,lb}, 1, \frac{rb\,-\,x}{rb\,-\,rt}), 0)$ on bottom and top vertices
$μ_{gaussian}(x) = e^{-\frac{1}{2}(\frac{x - t}{σ})^2}$ where σ is variance and t is the mean of the distribution
$μ_{bell}(x) = \frac{1}{1 + |1 + \frac{x - t}{a}|^{2b}}$ where l, t and r and left, top and right points of the curve
$μ_{sigmoid}(x,a,c)= \frac{1}{1 + e^{-a(x\,-\,c)}}$ where a is the width of transition area, c is the inflextion point
See also defuzz.

FuzzySystems.FuzzyLogic.Rule — Type

Rule(in::Tuple{Vararg{Symbol}}, out::Symbol, op::Union{Function, String, Symbol})

A structure storing fuzzy rule components.

Fuzzy rules map in to an outby&or|ops. Eachin`must be unique, and map to an unambiguousout`.

arg	alias

in	antecedent, premise, input

out	consequent, conclusion, output

The following are semantically equivalent

IF x₁ AND x₂ THEN y

y = x₁ & x₂

See also @rule, @rules, μ and firing