FuzzySystems.FuzzyLogic.defuzz — Function

defuzz(firing_strength, fis, method)

Obtain a crisp value from degrees of memberships of $x$.

universe of discourse 𝒰, the range of interest. mf the membership function method one of

Center of Gravity (Centroid) COG

Bisector of Area BOA

Weighted Average WAM

First of Maxima FOM

Last of Maxima LOM

Mean of Maxima MOM

Smallest of Maxima SOM

See also μ.

FuzzySystems.FuzzyLogic.implicationproperties — Method

Axiom adherence of an implication function

N is a strong negation function, i.e. a fuzzy complement ~~x == x. T is a left-continuous t-norm, defaults to nilpotent minimum

Mas, M., Monserrat, M., Torrens, J., & Trillas, E. (2007). A survey on fuzzy implication functions. IEEE Transactions on fuzzy systems, 15(6), 1107-1121.

FuzzySystems.FuzzyLogic.issnorm — Method

is a fuzzy union (AND) of the form [0,1]² -> [0,1] 1) associative, 2) monotone, 3) communicative and 4) bounded?

FuzzySystems.FuzzyLogic.istnorm — Method

is a fuzzy intersection (OR) function of the form [0,1]² -> [0,1] 1) associative, 2) monotone, 3) communicative and 4) bounded?

FuzzySystems.FuzzyLogic.logic — Method

Define fuzzy set operations logic(model, args...)

Each logic is made up of

FuzzySystems.FuzzyLogic.μ — Function

μ(x, mf::membership_function)

Obtain degrees of membership $μ$ for a given crisp input $x$

$μ_{triangular}(x)= max(min(\frac{x\,-\,l}{t\,-\,l}, \frac{r\,-\,x}{t\,-\,r}), 0)$, on left, top and right vertices
$μ_{trapezoid}(x)= max(min(\frac{x\,-\,lb}{lt\,-\,lb}, 1, \frac{rb\,-\,x}{rb\,-\,rt}), 0)$ on bottom and top vertices
$μ_{gaussian}(x) = e^{-\frac{1}{2}(\frac{x - t}{σ})^2}$ where σ is variance and t is the mean of the distribution
$μ_{bell}(x) = \frac{1}{1 + |1 + \frac{x - t}{a}|^{2b}}$ where l, t and r and left, top and right points of the curve
$μ_{sigmoid}(x,a,c)= \frac{1}{1 + e^{-a(x\,-\,c)}}$ where a is the width of transition area, c is the inflextion point
See also defuzz.